backend/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "inventory-backend"
version = "1.0.0"
description = "Inventory Management System Backend"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "sqlalchemy>=2.0.25",
    "alembic>=1.13.1",
    "pydantic[email]>=2.5.3",
    "pydantic-settings>=2.1.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib>=1.7.4",
    "bcrypt>=3.2.0,<4.0.0",
    "python-multipart>=0.0.6",
    "aiofiles>=23.2.1",
    "httpx>=0.26.0",
    "pandas>=2.1.4",
    "openpyxl>=3.1.2",
    "reportlab>=4.0.8",
    "pillow>=10.2.0",
    "onnxruntime>=1.16.3",
    "numpy>=1.26.3",
    "pytz>=2024.1",
    "typing_extensions>=4.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.4",
    "pytest-asyncio>=0.23.3",
    "pytest-cov>=4.1.0",
    "httpx>=0.26.0",
    "black>=24.1.0",
    "ruff>=0.1.14",
    "mypy>=1.8.0",
]
ai = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "easyocr>=1.7.0",
    "opencv-python>=4.8.0",
    "sentence-transformers>=2.2.0",  # Semantic text matching
]
ml = [
    # ML V2 Pipeline dependencies
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "ultralytics>=8.0.0",          # YOLOv8 for detection
    "timm>=0.9.0",                  # EfficientNet backbone
    "faiss-cpu>=1.7.4",             # Fast similarity search
    "onnx>=1.14.0",                 # Model export
    "onnxruntime>=1.16.0",          # ONNX inference
    "opencv-python>=4.8.0",
    "albumentations>=1.3.0",        # Data augmentation
    "scikit-learn>=1.3.0",          # Metrics
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
python_files = "test_*.py"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "W"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_ignores = true
